Finished task3

+++ task2.cpp
@@ -1,9 +1,10 @@
 #include <iostream>
 
-void print_size(const char* name, long size){
+void print_size(const char* name, size_t size){
     if (size == 1){
         std::cout << name << ": " << size << " byte" << std::endl;
     }
+    // Я не уверен, что это байты, но ладно, похоже на правду. На самом деле размер зависит от системы и size_t может иметь разные значения.
     else {
         std::cout << name << ": " << size << " bytes" << std::endl;
     }
